+++ components/okta/src/triggers/updatedUsersPollingTrigger.ts
@@ -16,13 +16,9 @@ export const updatedUsersPollingTrigger = pollingTrigger({
       : undefined;
     const params = {
       connection,
-      after: undefined,
-      limit: undefined,
-      q: undefined,
-      search: undefined,
-      sortBy: undefined,
-      sortOrder: undefined,
-      filter,
+      pagination: { after: undefined, limit: undefined },
+      sorting: { sortBy: undefined, sortOrder: undefined },
+      filters: { search: undefined, filter, q: undefined },
       fetchAll: true,
       extraParameters: {},
     };